--- app/src/main/java/com/imhungry/jjongseol/data/model/home/SuccessResponse.kt
@@ -1,7 +1,0 @@
-package com.imhungry.jjongseol.data.model.home
-
-data class SuccessResponse<T>(
-    val status: Int,
-    val message: String,
-    val data: T
-)
