+++ app/src/main/java/com/imhungry/jjongseol/data/model/response/SuccessResponse.kt
@@ -4,3 +4,9 @@ data class SuccessResponse<T>(
     val message: String,
     val data: T
 )
+
+/*data class SuccessResponse<T>(
+    val status: Int,
+    val message: String,
+    val data: T
+)*/
